move repl renderer related stuff to the replViewer
